combine state and spec for a resource
Due to the complexity involved in staging changes,
the specification has been corrected to combine spec
and state into one.
There will now be a single definition for a resource spec
that can have user-defined and computed fields,
computed fields will be configuration generated after
deploying a resource such as an identifier.